+++ Underperforming-Students-vs-Teachers/cgagod.cpp
@@ -3,7 +3,7 @@
 const QString CgaGod::DEFAULT_NAME = "CGA God";
 
 
-CgaGod::CgaGod(QLabel* widget, Row* row, QString name, int maxHp, double skillSpeed, int cost, int damage)
+CgaGod::CgaGod(QLabel* widget, Row* row, QString name, int maxHp, int skillSpeed, int cost, int damage)
     : AttackStudent(widget, row, name, maxHp, skillSpeed, cost, damage)
 {
