+++ Underperforming-Students-vs-Teachers/kelvin.cpp
@@ -3,7 +3,7 @@
 const QString Kelvin::DEFAULT_NAME = "Kelvin";
 
 
-Kelvin::Kelvin(QLabel *widget, Row* row, QString name, int maxHp, double speed, int damage) :
+Kelvin::Kelvin(QLabel *widget, Row* row, QString name, int maxHp, int speed, int damage) :
     Teacher(widget, row, name, maxHp, speed, damage)
 {
