+++ Underperforming-Students-vs-Teachers/supportstudent.cpp
@@ -1,6 +1,6 @@
 #include "supportstudent.h"
 
-SupportStudent::SupportStudent(QLabel* widget, Row* row, QString name, int maxHp, double skillSpeed, int cost)
+SupportStudent::SupportStudent(QLabel* widget, Row* row, QString name, int maxHp, int skillSpeed, int cost)
     : Student(widget, row, name, maxHp, skillSpeed, cost)
 {
